与标准库一致

+++ lib/types.h
@@ -4,12 +4,12 @@
 typedef unsigned char uint8_t;
 typedef unsigned short uint16_t;
 typedef unsigned uint32_t;
-typedef unsigned long long uint64_t;
+typedef unsigned long uint64_t;
 
 typedef char int8_t;
 typedef short int16_t;
 typedef int int32_t;
-typedef long long int64_t;
+typedef long int64_t;
 
 typedef uint64_t size_t;
 typedef int64_t ssize_t;
